requirements.txt
# local package
-e .

pyaudio
wave
git+https://github.com/openai/whisper.git
pydub
pandas
python-dotenv
pyannote.audio

# external requirements
click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1
